+++ stream-chat-android-docs/src/main/java/io/getstream/chat/docs/java/Android.java
@@ -910,7 +910,8 @@ public class Android {
                                 "Empty state text",
                                 genericDrawable,
                                 textStyleGeneric,
-                                genericDrawable
+                                genericDrawable,
+                                10
                         )
 
             );